FeynmanDiagram

Documentation for FeynmanDiagram.

FeynmanDiagram.ComputationalGraphs.FeynmanGraph

FeynmanDiagram.ComputationalGraphs.Graph

FeynmanDiagram.DiagTree.BareGreenNId

FeynmanDiagram.DiagTree.BareHoppingId

FeynmanDiagram.DiagTree.ConnectedGreenNId

FeynmanDiagram.DiagTree.Diagram

FeynmanDiagram.DiagTree.DiagramId

FeynmanDiagram.DiagTree.GreenNId

FeynmanDiagram.DiagTree.PropagatorId

FeynmanDiagram.ExprTree.LoopPool

FeynmanDiagram.ExprTree.Node

FeynmanDiagram.FrontEnds.LabelProduct

FeynmanDiagram.FrontEnds.LoopPool

FeynmanDiagram.Parquet.ParquetBlocks

FeynmanDiagram.QuantumOperators.OperatorProduct

FeynmanDiagram.QuantumOperators.QuantumOperator

Base.:*

Base.:*

Base.:*

Base.:*

Base.:*

Base.:*

Base.:+

Base.:+

Base.:-

Base.:-

Base.adjoint

Base.adjoint

Base.convert

Base.length

Base.show

Base.show

Base.size

Base.size

FeynmanDiagram.ComputationalGraphs.alleq

FeynmanDiagram.ComputationalGraphs.build_derivative_graph

FeynmanDiagram.ComputationalGraphs.burn_from_targetleaves!

FeynmanDiagram.ComputationalGraphs.collect_labels

FeynmanDiagram.ComputationalGraphs.constant_graph

FeynmanDiagram.ComputationalGraphs.count_operation

FeynmanDiagram.ComputationalGraphs.diagram_type

FeynmanDiagram.ComputationalGraphs.drop_topology

FeynmanDiagram.ComputationalGraphs.eldest

FeynmanDiagram.ComputationalGraphs.external_indices

FeynmanDiagram.ComputationalGraphs.external_labels

FeynmanDiagram.ComputationalGraphs.external_legs

FeynmanDiagram.ComputationalGraphs.external_operators

FeynmanDiagram.ComputationalGraphs.external_vertex

FeynmanDiagram.ComputationalGraphs.factor

FeynmanDiagram.ComputationalGraphs.feynman_diagram

FeynmanDiagram.ComputationalGraphs.flatten_all_chains!

FeynmanDiagram.ComputationalGraphs.flatten_all_chains!

FeynmanDiagram.ComputationalGraphs.flatten_chains

FeynmanDiagram.ComputationalGraphs.flatten_chains!

FeynmanDiagram.ComputationalGraphs.forwardAD

FeynmanDiagram.ComputationalGraphs.forwardAD_root!

FeynmanDiagram.ComputationalGraphs.group

FeynmanDiagram.ComputationalGraphs.haschildren

FeynmanDiagram.ComputationalGraphs.id

FeynmanDiagram.ComputationalGraphs.interaction

FeynmanDiagram.ComputationalGraphs.is_external

FeynmanDiagram.ComputationalGraphs.is_internal

FeynmanDiagram.ComputationalGraphs.isapprox_one

FeynmanDiagram.ComputationalGraphs.isassociative

FeynmanDiagram.ComputationalGraphs.isbranch

FeynmanDiagram.ComputationalGraphs.ischain

FeynmanDiagram.ComputationalGraphs.isequiv

FeynmanDiagram.ComputationalGraphs.isfactorless

FeynmanDiagram.ComputationalGraphs.isleaf

FeynmanDiagram.ComputationalGraphs.linear_combination

FeynmanDiagram.ComputationalGraphs.linear_combination

FeynmanDiagram.ComputationalGraphs.linear_combination

FeynmanDiagram.ComputationalGraphs.linear_combination

FeynmanDiagram.ComputationalGraphs.merge_all_linear_combinations!

FeynmanDiagram.ComputationalGraphs.merge_all_linear_combinations!

FeynmanDiagram.ComputationalGraphs.merge_all_multi_products!

FeynmanDiagram.ComputationalGraphs.merge_all_multi_products!

FeynmanDiagram.ComputationalGraphs.merge_linear_combination

FeynmanDiagram.ComputationalGraphs.merge_linear_combination!

FeynmanDiagram.ComputationalGraphs.merge_multi_product

FeynmanDiagram.ComputationalGraphs.merge_multi_product!

FeynmanDiagram.ComputationalGraphs.multi_product

FeynmanDiagram.ComputationalGraphs.multi_product

FeynmanDiagram.ComputationalGraphs.name

FeynmanDiagram.ComputationalGraphs.onechild

FeynmanDiagram.ComputationalGraphs.operator

FeynmanDiagram.ComputationalGraphs.optimize

FeynmanDiagram.ComputationalGraphs.optimize!

FeynmanDiagram.ComputationalGraphs.orders

FeynmanDiagram.ComputationalGraphs.plot_tree

FeynmanDiagram.ComputationalGraphs.propagator

FeynmanDiagram.ComputationalGraphs.relabel

FeynmanDiagram.ComputationalGraphs.relabel!

FeynmanDiagram.ComputationalGraphs.remove_duplicated_leaves!

FeynmanDiagram.ComputationalGraphs.replace_subgraph

FeynmanDiagram.ComputationalGraphs.replace_subgraph!

FeynmanDiagram.ComputationalGraphs.set_factor!

FeynmanDiagram.ComputationalGraphs.set_id!

FeynmanDiagram.ComputationalGraphs.set_name!

FeynmanDiagram.ComputationalGraphs.set_operator!

FeynmanDiagram.ComputationalGraphs.set_operator!

FeynmanDiagram.ComputationalGraphs.set_orders!

FeynmanDiagram.ComputationalGraphs.set_subgraph!

FeynmanDiagram.ComputationalGraphs.set_subgraph_factor!

FeynmanDiagram.ComputationalGraphs.set_subgraph_factors!

FeynmanDiagram.ComputationalGraphs.set_subgraph_factors!

FeynmanDiagram.ComputationalGraphs.set_subgraphs!

FeynmanDiagram.ComputationalGraphs.set_subgraphs!

FeynmanDiagram.ComputationalGraphs.set_weight!

FeynmanDiagram.ComputationalGraphs.standardize_labels

FeynmanDiagram.ComputationalGraphs.standardize_labels!

FeynmanDiagram.ComputationalGraphs.subgraph

FeynmanDiagram.ComputationalGraphs.subgraph_factor

FeynmanDiagram.ComputationalGraphs.subgraph_factors

FeynmanDiagram.ComputationalGraphs.subgraph_factors

FeynmanDiagram.ComputationalGraphs.subgraphs

FeynmanDiagram.ComputationalGraphs.subgraphs

FeynmanDiagram.ComputationalGraphs.topology

FeynmanDiagram.ComputationalGraphs.unary_istrivial

FeynmanDiagram.ComputationalGraphs.unique_leaves

FeynmanDiagram.ComputationalGraphs.vertex

FeynmanDiagram.ComputationalGraphs.vertices

FeynmanDiagram.ComputationalGraphs.weight

FeynmanDiagram.DiagTree.derivative

FeynmanDiagram.DiagTree.derivative

FeynmanDiagram.DiagTree.plot_tree

FeynmanDiagram.DiagTree.removeDuplicatedLeaves!

FeynmanDiagram.DiagTree.removeHartreeFock!

FeynmanDiagram.DiagTree.removeOneChildParent!

FeynmanDiagram.ExprTree.addnode!

FeynmanDiagram.ExprTree.addpropagator!

FeynmanDiagram.ExprTree.getNode

FeynmanDiagram.ExprTree.getNodeWeight

FeynmanDiagram.ExprTree.showTree

FeynmanDiagram.FrontEnds.Graph!

FeynmanDiagram.FrontEnds.extract_var_dependence

FeynmanDiagram.FrontEnds.index_to_linear

FeynmanDiagram.FrontEnds.linear_to_index

FeynmanDiagram.GV.diagdictGV

FeynmanDiagram.GV.diagdictGV

FeynmanDiagram.GV.eachorder_diag

FeynmanDiagram.GV.leafstates

FeynmanDiagram.GV.read_diagrams

FeynmanDiagram.Parquet.ep_coupling

FeynmanDiagram.Parquet.green

FeynmanDiagram.Parquet.polarization

FeynmanDiagram.Parquet.sigma

FeynmanDiagram.Parquet.vertex3

FeynmanDiagram.Parquet.vertex4

FeynmanDiagram.QuantumOperators.correlator_order

FeynmanDiagram.QuantumOperators.fermionic_annihilation

FeynmanDiagram.QuantumOperators.isannihilation

FeynmanDiagram.QuantumOperators.iscreation

FeynmanDiagram.QuantumOperators.isfermionic

FeynmanDiagram.QuantumOperators.isfermionic

FeynmanDiagram.QuantumOperators.normal_order

FeynmanDiagram.QuantumOperators.parity

FeynmanDiagram.QuantumOperators.parity_old

FeynmanDiagram.innerTauNum

FeynmanDiagram.innerTauNum

Parameters.reconstruct

Parameters.reconstruct

FeynmanDiagram.innerTauNum — Method

function innerTauNum(type, innerLoopNum, interactionTauNum)

internal imaginary-time degrees of freedom for a given diagram type and internal loop number.
For the vertex functions (self-energy, polarization, vertex3, and vertex4), innerTauNum is equivalent to tauNum.
For the Green function, tauNum = innerTauNum + external tauNum

FeynmanDiagram.innerTauNum — Method

function innerTauNum(type::DiagramType, innerLoopNum, interactionTauNum)

internal imaginary-time degrees of freedom for a given diagram type and internal loop number.
For the vertex functions (self-energy, polarization, vertex3, and vertex4), innerTauNum is equivalent to tauNum.
For the Green function, tauNum = innerTauNum + external tauNum

Parameters.reconstruct — Method

Parameters.reconstruct(p::DiagPara; kws...)

Type-stable version of the Parameters.reconstruct

Parameters.reconstruct — Method

Parameters.reconstruct(p::DiagramPara; kws...)

Type-stable version of the Parameters.reconstruct

Library Outline

Quantum operator and its product

API

Computational graph for general feynman diagrams

API

Front-end translates a source code into a computational graph

API

GV expansion to a computational graph

API

Parquet Algorithm to Build Diagrams

API

Usage

Diagrams as an AbstractTree

API

Diagrams as an Expression Tree

API