GV expansion to a computational graph

API

FeynmanDiagram.GV.diagdictGV — Function

function diagdictGV(type::Symbol, MaxOrder::Int, has_counterterm::Bool=false, dim::Int=3)

Generates a Graph Dict: the `dim`-dimensional spin/charge polarization or self-energy diagrams with static interactions in a given `type`, to a given maximum order `MaxOrder`, with switchable couterterms. 
Generates fermionic/bosonic `LabelProduct`: `fermi_labelProd`/`bose_labelProd` for these Graphs.
Generates a Tuple (propagatorMap, interactionMap) for mapping `g.id` to the index of unique proapgators and interactions, respectively.

Arguments:

type (Symbol): The type of the Feynman diagrams, including :spinPolar, :chargePolar, :sigma, :green, or :freeEnergy.

gkeys (Vector{Tuple{Int,Int,Int}}): The (order, Gorder, Vorder) of the diagrams. Gorder is the order of self-energy counterterms, and Vorder is the order of interaction counterterms.

dim (Int): The dimension of the system (defaults to 3).

spinPolarPara (Float64, optional): The spin-polarization parameter (nup - ndown) / (nup + ndown) (defaults to 0.0).

Returns

A tuple (diagrams, fermi_labelProd, bose_labelProd) where

diagrams is a Dict{Tuple{Int,Int,Int},Tuple{Vector{Graph},Vector{Vector{Int}}}} object representing the diagrams. The key is (order, Gorder, Vorder). The element is a Tuple (diagrams, extT_labels).

fermi_labelProd is a LabelProduct object containing the labels for the fermionic G objects in the diagrams,

bose_labelProd is a LabelProduct object containing the labels for the bosonic W objects in the diagrams.

(propagatorMap, interactionMap) maps g.id to the index of unique proapgators and interactions, respectively.

FeynmanDiagram.GV.diagdictGV — Function

function diagdictGV(type::Symbol, MaxOrder::Int, has_counterterm::Bool=false, dim::Int=3)

Generates a Graph Dict: the `dim`-dimensional spin/charge polarization or self-energy diagrams with static interactions in a given `type`, to a given maximum order `MaxOrder`, with switchable couterterms. 
Generates fermionic/bosonic `LabelProduct`: `fermi_labelProd`/`bose_labelProd` for these Graphs.
Generates a Tuple (propagatorMap, interactionMap) for mapping `g.id` to the index of unique proapgators and interactions, respectively.

Arguments:

type (Symbol): The type of the Feynman diagrams, including :spinPolar, :chargePolar, :sigma, :green, or :freeEnergy.

Maxorder (Int): The maximum actual order of the diagrams.

has_counterterm (Bool): false for G0W0, true for GW with self-energy and interaction counterterms (defaults to false).

dim (Int): The dimension of the system (defaults to 3).

spinPolarPara (Float64, optional): The spin-polarization parameter (nup - ndown) / (nup + ndown) (defaults to 0.0).

Returns

A tuple (diagrams, fermi_labelProd, bose_labelProd) where

diagrams is a Dict{Tuple{Int,Int,Int},Tuple{Vector{Graph},Vector{Vector{Int}}}} object representing the diagrams. The key is (order, Gorder, Vorder). The element is a Tuple (diagrams, extT_labels).

fermi_labelProd is a LabelProduct object containing the labels for the fermionic G objects in the diagrams,

bose_labelProd is a LabelProduct object containing the labels for the bosonic W objects in the diagrams.

(propagatorMap, interactionMap) maps g.id to the index of unique proapgators and interactions, respectively.

FeynmanDiagram.GV.eachorder_diag — Function

function eachorder_diag(type::Symbol, order::Int, VerOrder::Int=0, GOrder::Int=0; loopPool::Union{LoopPool,Nothing}=nothing,
    tau_labels::Union{Nothing,Vector{Int}}=nothing, GTypes::Union{Nothing,Vector{Int}}=nothing, VTypes::Union{Nothing,Vector{Int}}=nothing)

Generates a `Vector{Graph}`: the polarization diagrams with static interactions of a given order, where the actual order of diagrams equals to `order + VerOrder + 2 * GOrder`.
Generates fermionic/bosonic `LabelProduct`: `fermi_labelProd`/`bose_labelProd` with inputs `tau_labels`, `GTypes`/`VTypes`, and updated `loopPool`. 
Generates external tau labels Vector{Vector{Int}}. The i-th labels (Vector{Int}) corresponds to the i-th `Graph` in `Vector{Graph}`.

Arguments:

type (Symbol): The type of the diagrams, including :spinPolar, :chargePolar, :sigma, :green, or :freeEnergy.

order (Int): The order of the diagrams without counterterms.

VerOrder (Int, optional): The order of interaction counterterms (defaults to 0).

GOrder (Int, optional): The order of self-energy counterterms (defaults to 0).

dim (Int, optional): The dimension of the system (defaults to 3).

spinPolarPara (Float64, optional): The spin-polarization parameter (nup - ndown) / (nup + ndown) (defaults to 0.0).

loopPool (Union{LoopPool,Nothing}=nothing, optional): The initial pool of loop momenta (defaults to nothing).

tau_labels(Union{Nothing, Vector{Int}}, optional): The labels for the discrete time of each vertex. (defaults to nothing).

GTypes: The types of fermion propagators G in the diagrams (defaults to collect(0:GOrder)).

VTypes: The types of boson static interaction V in the diagrams (defaults to collect(0:VerOrder)).

Returns

A tuple (diagrams, fermi_labelProd, bose_labelProd, extT_labels) where

diagrams is a Vector{Graph} object representing the diagrams,

fermi_labelProd is a LabelProduct object containing the labels for the fermionic G objects in the diagrams,

bose_labelProd is a LabelProduct object containing the labels for the bosonic W objects in the diagrams.

extT_labels is a Vector{Vector{Int}} object containing the external tau labels for each Graph in diagrams.

FeynmanDiagram.GV.leafstates — Method

function leafstates(
    FeynGraphs::Dict{T, Tuple{Vector{G}, Vector{Vector{Int}}}},
    FermiLabel::LabelProduct, BoseLabel::LabelProduct,
    graph_keys::Vector{T}
) where {T, G <: Graph}

Extracts leaf information from a Dict collection of Feynman graphs (`FeynGraphs` with its keys `graph_keys`)
and their associated LabelProduct data (`FermiLabel` and `BoseLabel`). 
The information includes their initial value, type, in/out time, and loop momenta.

Arguments:

FeynGraphs: A dictionary mapping keys of type T to tuples containing a vector of Graph objects and a vector of external time labels.

FermiLabel: A LabelProduct used to label the fermionic G objects in the graphs.

BoseLabel: A LabelProduct used to label bosonic W objects in the graphs.

graph_keys: A vector containing keys of type T, specifying which graphs to analyze.

Returns

A tuple of vectors containing information about the propagators in the graphs, including their initial values, types, input and output time indexes, and loop-momenta indexes.

A tuple of vectors containing information about the interactions in the graphs, including their initial values, types, input and output time indexes, and loop-momenta indexes.

A Vector{Vector{Int}} representing the external tau variables of each vector of graph corresponding to each key of type T.

FeynmanDiagram.GV.read_diagrams — Method

function read_diagrams(filename::AbstractString; loopPool::Union{LoopPool,Nothing}=nothing,
    dim::Int=3, tau_labels::Union{Nothing,Vector{Int}}=nothing, GTypes=[0, 1], VTypes=[0, 1, 2],
    keywords::Vector{String}=["Polarization", "DiagNum", "Order", "GNum", "Ver4Num", "LoopNum", "ExtLoopIndex",
        "DummyLoopIndex", "TauNum", "ExtTauIndex", "DummyTauIndex"])

Reads a GV_diagrams file and returns Graph of diagrams in this file 
and the corresponding `LabelProduct` objects, which are used to keep track of QuantumOperator.label.

Arguments:

filename (AbstractString): The path to the file containing the diagrams.

loopPool (Union{LoopPool,Nothing}): An optional LoopPool object. If not provided, a new one will be created.

spinPolarPara (Float64): The spin-polarization parameter (nup - ndown) / (nup + ndown) (defaults to 0.0).

dim (Int): The dimension of the system, used to initialize the LoopPool object. Default is 3.

tau_labels (Union{Nothing,Vector{Int}}): The labels for the Tau objects in the diagrams. If not provided, they will be set to the integers from 1 to tauNum.

GTypes (Vector{Int}): The labels for the fermionic G objects in the diagrams. Default is [0, 1].

VTypes (Vector{Int}): The labels for the bosonic V objects in the diagrams. Default is [0, 1, 2].

keywords (Vector{String}): A set of keywords used to extract information from the file. Default is ["Polarization", "DiagNum", "Order", "GNum", "Ver4Num", "LoopNum", "ExtLoopIndex", "DummyLoopIndex", "TauNum", "ExtTauIndex", "DummyTauIndex"].

Returns

A tuple (diagrams, fermi_labelProd, bose_labelProd) where

diagrams is a Graph object representing the diagrams,

fermi_labelProd is a LabelProduct object containing the labels for the fermionic G objects in the diagrams,

bose_labelProd is a LabelProduct object containing the labels for the bosonic W objects in the diagrams.