Diagrams as an AbstractTree

API

FeynmanDiagram.DiagTree.BareGreenNId — Type

time-ordered N-point Bare Green's function

FeynmanDiagram.DiagTree.BareHoppingId — Type

hopping function c⁺c⁻

FeynmanDiagram.DiagTree.ConnectedGreenNId — Type

time-ordered N-point Composite Green's function

FeynmanDiagram.DiagTree.Diagram — Type

mutable struct Diagram{W}

struct of a diagram. A diagram of a sum or produce of various subdiagrams.

Members

hash::Int : the unique hash number to identify the diagram

name::Symbol : name of the diagram

id::DiagramId : diagram id

operator::Operator : operation, support Sum() and Prod()

factor::W : additional factor of the diagram

subdiagram::Vector{Diagram{W}} : vector of sub-diagrams

weight::W : weight of the diagram

FeynmanDiagram.DiagTree.DiagramId — Type

abstract type DiagramId end

The abstract type of all diagrams/subdiagrams/bare propagators

FeynmanDiagram.DiagTree.GreenNId — Type

time-ordered N-point Composite Green's function

FeynmanDiagram.DiagTree.PropagatorId — Type

abstract type PropagatorId <: DiagramId end

The abstract type of all bare propagators

FeynmanDiagram.DiagTree.derivative — Method

function derivative(diags::Union{Diagram,Tuple,AbstractVector}, ::Type{ID}, order::Int) where {ID<:PropagatorId}

Automatic differentiation derivative on the diagrams

Arguments

diags : diagrams to take derivative

ID : DiagramId to apply the differentiation

order::Int : derivative order

FeynmanDiagram.DiagTree.derivative — Method

function derivative(diags::Union{Tuple,AbstractVector}, ::Type{ID}; index::Int=index(ID)) where {W,ID<:PropagatorId}
function derivative(diags::Vector{Diagram{W}}, ::Type{ID}; index::Int=index(ID)) where {W,ID<:PropagatorId}

Automatic differentiation derivative on the diagrams

Arguments

diags : diagrams to take derivative

ID : DiagramId to apply the differentiation

index : index of the id.order array element to increase the order

FeynmanDiagram.DiagTree.plot_tree — Method

function plot_tree(diag::Diagram; verbose = 0, maxdepth = 6)

Visualize the diagram tree using ete3 python package

#Arguments

diag : the Diagram struct to visualize

verbose=0 : the amount of information to show

maxdepth=6 : deepest level of the diagram tree to show

FeynmanDiagram.DiagTree.removeDuplicatedLeaves! — Method

removeDuplicatedLeaves!(diags::AbstractVector; verbose = 0)

remove duplicated nodes such as:  ---> ver4 ---> InteractionId. Leaf will not be touched!

FeynmanDiagram.DiagTree.removeHartreeFock! — Method

function removeHartreeFock!(diag::Diagram{W}) where {W}
function removeHartreeFock!(diags::Union{Tuple,AbstractVector})

Remove the Hartree-Fock insertions that without any derivatives on the propagator and the interaction.

Arguments

diags : diagrams to remove the Fock insertion

Remarks

The operations removeHartreeFock! and taking derivatives doesn't commute with each other!

If the input diagram is a Hartree-Fock diagram, then the overall weight will become zero!

The return value is always nothing

FeynmanDiagram.DiagTree.removeOneChildParent! — Method

removeOneChildParent!(diags::AbstractVector; verbose = 0)

remove duplicated nodes such as:  ---> ver4 ---> InteractionId. Leaf will not be touched!