Diagrams as an Expression Tree

API

FeynmanDiagram.ExprTree.LoopPool — Type

struct LoopPool{T}

Pool of loop basis. Each loop basis corresponds to a loop variable.
A loop variable is a linear combination of N independent loops. The combination coefficients is what we call a loop basis.
For example, if a loop is a momentum K, then

varibale_i = K_1*basis[1, i] + K_2*basis[2, i] + K_3*basis[3, i] + ...

Members

name::Symbol : name of the pool

dim::Int : dimension of a loop variable (for example, the dimension of a momentum-frequency loop variable is (d+1) where d is the spatial dimension)

N::Int : number of independent loops (dimension of loop basis)

basis::Matrix{T} : Matrix of (N x Nb) that stores the loop basis, where Nb is the number of loop basis (or number of loop variables).

current::Matrix{T} : Matrix of (dim x Nb) that stores the loop variables, where Nb is the number of loop basis (or number of loop variables).

FeynmanDiagram.ExprTree.Node — Type

mutable struct Node{PARA,F}

Node Object, which is the building block of the diagram tree. Each node is a collection of CACHED proapgator objects and other child CACHED node objects

Members

para::PARA : user-defined parameters, which will be used to evaluate the factor and the weight of the node (e.g., if the node represents a vertex function, then the parameter may be the momentum basis of the external legs)

operation::Int : #1: multiply, 2: add, ...

factor::F : additional factor of the node

components::Vector{Vector{Int}} : Index to the cached propagators stored in certain pools. Each Vector{Int} is for one kind of propagator.

childNodes::Vector{Int} : Indices to the cached nodes stored in certain pool. They are the child of the current node in the diagram tree.

parent::Int : Index to the cached nodes which is the parent of the current node.

FeynmanDiagram.ExprTree.addnode! — Method

function addnode!(diag::ExpressionTree{V,PARA,F,W}, operator, name, children::Union{Tuple, AbstractVector}, factor = 1.0; para = nothing) where {V,PARA,F,W}

Add a node into the expression tree.

Arguments

diag::ExpressionTree : diagrammatic experssion tree.

operator::Int : #1: multiply, 2: add, ...

name : name of the node

children : Indices to the cached nodes stored in certain pool. They are the child of the current node in the diagram tree. It should be in the format of Vector{Int}.

factor = 1.0 : Factor of the node

para = nothing : Additional paramenter required to evaluate the node. Set to nothing by default.

FeynmanDiagram.ExprTree.addpropagator! — Method

function addPropagator!(diag::ExpressionTree, name, factor = 1.0; site = [], loop = nothing, para = nothing, order::Int = 0)

Add a propagator into the diagram tree.

Arguments

diag : diagrammatic experssion tree.

order::Int = 0 : Order of the propagator.

name = :none : name of the propagator.

factor = 1 : Factor of the propagator.

site = [] : site basis (e.g, time and space coordinate) of the propagator.

loop = nothing : loop basis (e.g, momentum and frequency) of the propagator.

para = nothing : Additional paramenter required to evaluate the propagator.

FeynmanDiagram.ExprTree.getNode — Method

function getNode(diag::Diagrams, nidx::Int)

get Node in the diag with the index nidx.

FeynmanDiagram.ExprTree.getNodeWeight — Method

function getNodeWeight(tree, nidx::Int)

get Node weight in the diagram experssion tree with the index nidx.

FeynmanDiagram.ExprTree.showTree — Method

showTree(diag::Diagrams, _root = diag.root[end]; verbose = 0, depth = 999)

Visualize the diagram tree using ete3 python package

#Arguments

diag: the Diagrams struct to visualize

_root: the index of the root node to visualize

verbose=0: the amount of information to show

depth=999: deepest level of the diagram tree to show