Front-end translates a source code into a computational graph

API

FeynmanDiagram.FrontEnds.LabelProduct — Type

The cartesian QuantumOperator.label product:

#Parameters:

'LT': Type of labels

'N' : Number of labels' type

#Members:

'labels' : The list of labels in the LabelProduct

'dims' : A tuple of the length of the label factors

FeynmanDiagram.FrontEnds.LoopPool — Type

struct LoopPool{T}

Pool of loop basis. Each loop basis corresponds to a loop variable.
A loop variable is a linear combination of N independent loops. The combination coefficients is what we call a loop basis.
For example, if a loop is a momentum K, then

varibale_i = K_1*basis[1, i] + K_2*basis[2, i] + K_3*basis[3, i] + ...

Members

name::Symbol : name of the pool

dim::Int : dimension of a loop variable (for example, the dimension of a momentum-frequency loop variable is (d+1) where d is the spatial dimension)

basis::Matrix{T} : Matrix of (N x Nb) that stores the loop basis, where Nb is the number of loop basis (or number of loop variables).

loops::Matrix{T} : Matrix of (dim x Nb) that stores the loop variables, where Nb is the number of loop basis (or number of loop variables).

Base.length — Method

function Base.length(obj::LabelProduct)

Return the number of grids of the LabelProduct.

Base.show — Method

function Base.show(io::IO, obj::LabelProduct)

Print the LabelProduct.

Base.size — Method

function Base.size(obj::LabelProduct, I::Int)

Return the length of the specifict Ith label factor of the LabelProduct.

Base.size — Method

function Base.size(obj::LabelProduct, I::Int)

Return the length of the specifict Ith label factor of the LabelProduct.

FeynmanDiagram.FrontEnds.index_to_linear — Method

function index_to_linear(obj::LabelProduct, index...)

Convert a tuple of the indexes of each label to a single linear index of the LabelProduct.

Argument:

'obj': The LabelProduct object

'index...': N indexes of the label factor, where N is the number of label factor

FeynmanDiagram.FrontEnds.linear_to_index — Method

function linear_to_index(obj::LabelProduct, I::Int)

Convert the single linear index of the LabelProduct to a tuple of indexes of each label.

Argument:

'obj': The LabelProduct object

'I': The linear index of the LabelProduct