Diagram Tree Builder based on the Parquet Algorithm

FeynmanDiagram.Builder.Parquet.IdxMap — Type

struct IdxMap{_Ver4}

Map left vertex Tpair[lidx], right vertex Tpair[ridx], the shared Green's function G0 and the channel specific Green's function Gx to the top level 4-vertex Tpair[vidx]

Arguments

l::_Ver4 : left vertex

r::_Ver4 : right vertex

v::_Ver4 : composte vertex

lidx::Int : left sub-vertex index

ridx::Int : right sub-vertex index

vidx::Int : composite vertex index

G0::Green : shared Green's function index

Gx::Green : channel specific Green's function index

FeynmanDiagram.Builder.Parquet.Ver4 — Type

function Ver4{W}(para::Para, loopNum = para.internalLoopNum, tidx = 1; chan = para.chan, F = para.F, V = para.V, level = 1, id = [1,]) where {W}

Generate 4-vertex diagrams using Parquet Algorithm

#Arguments

para: parameters. It should provide internalLoopNum, interactionTauNum, firstTauIdx

chan: list of channels of the current 4-vertex.

F : channels of left sub-vertex for the particle-hole and particle-hole-exchange bubbles

V : channels of left sub-vertex for the particle-particle bubble

All : channels of right sub-vertex of all channels

Fouter : channels of left sub-vertex for the particle-hole and particle-hole-exchange bubbles, only take effect for the outermost bubble

Vouter : channels of left sub-vertex for the particle-particle bubble, only take effect for the outermost bubble

Allouter : channels of right sub-vertex of all channels

loopNum: momentum loop degrees of freedom of the 4-vertex diagrams

tidx: the first τ variable index. It will be the τ variable of the left incoming electron for all 4-vertex diagrams

level: level in the diagram tree

id: the first element will be used as the id of the Ver4. All nodes in the tree will be labeled in preorder depth-first search

#Remark:

AbstractTrees interface is implemented for Ver4. So one can use the API in https://juliacollections.github.io/AbstractTrees.jl/stable/ to manipulate/print the tree structre of Ver4.

There are three different methods to print/visualize the tree structre:

print_tree(ver4::Ver4) or print_tree(bub::Bubble) to print the tree to terminal. This function is provided by AbstractTrees API.

newick(ver4::Ver4) or newick(bub::Bubble) to serilize the tree to a newick format string. You may save the string to a text file, then visualize it with a newick format visualizer application.

showTree(ver4::Ver4) to visualize the tree using the python package ete3. You have to install ete3 properly to use this function.

FeynmanDiagram.Builder.Parquet.buildG — Function

function buildG(para, externLoop, extT, subdiagram = false; F = [I, U, S], V = [I, T, U], All = union(F, V), diag = newDiagTree(para, :G))

Build composite Green's function.
By definition, para.firstTauIdx is the first Tau index of the left most self-energy subdiagram.

extT: [Tau index of the left leg, Tau index of the right leg]

FeynmanDiagram.Builder.Parquet.buildSigma — Function

function buildSigma(para, externLoop; F = [I, U, S], V = [I, T, U], All = union(F, V), diag = newDiagTree(para, :Sigma), subdiagram = false)

build sigma diagram. 
When sigma is created as a subdiagram, then no Fock diagram is generated if para.filter contains NoFock, and no sigma diagram is generated if para.filter contains Girreducible

FeynmanDiagram.Builder.Parquet.buildVer4 — Function

function buildVer4(para, LegK, chan, F, V, All = union(F, V);
    Fouter = F, Vouter = V, Allouter = All, factor = 1.0, diag = newDiagTree(para, Tuple{Int,Int,Int,Int}, :Ver4))

build DiagTree for the one-particle-irreducible 4-point vertex function using the parquet algorithm

Arguments:

weightType : type of the weight of the propagators and the vertex functions

para : parameters to generate the diagram tree

LegK : momentum basis of external legs, only three of them are expected: [left in, left out, right in], the dimension of each legK is called loopBasis dimension.

FeynmanDiagram.Builder.Parquet.newick — Method

convert Bubble tree struct to a string in the newick format

FeynmanDiagram.Builder.Parquet.newick — Method

convert Ver4 tree struct to a string in the newick format

FeynmanDiagram.Builder.Parquet.showTree — Method

showTree(ver4, para::Para; verbose=0, depth=999)

Visualize the diagram tree using ete3 python package

#Arguments

ver4: the 4-vertex diagram tree to visualize

para: parameters

verbose=0: the amount of information to show

depth=999: deepest level of the diagram tree to show