Parquet algorithm to generate diagram tree

ExpressionTree.Parquet.IdxMap — Type

struct IdxMap(lv, rv, G0, Gx, ver)

Map left vertex Tpair[lv], right vertex Tpair[rv], the shared Green's function G0[G0] and the channel specific Green's function Gx[Gx] to the top level 4-vertex Tpair[ver]

ExpressionTree.Parquet.Para — Type

Para(chan, interactionTauNum)

Parameters to generate diagrams using Parquet algorithm

#Arguments

chan: list of channels of sub-vertices

interactionTauNum: τ degrees of freedom of the bare interaction

ExpressionTree.Parquet.Ver4 — Type

Ver4{W}(loopNum, tidx, para::Para; chan=nothing, level=1, id=[1, ]) where {W}

Generate 4-vertex diagrams using Parquet Algorithm

#Arguments

loopNum: momentum loop degrees of freedom of the 4-vertex diagrams

tidx: the first τ variable index. It is also the τ variable of the left incoming electron for all 4-vertex diagrams

para: parameters

chan: list of channels of the current 4-vertex. If not specified, it is set to be para.chan

level: level in the diagram tree

id: the first element will be used as the id of the Ver4. All nodes in the tree will be labeled in preorder depth-first search

#Remark:

The argument chan and para.chan are different. The former is the channels of current 4-vertex, while the later is the channels of the sub-vertices

AbstractTrees interface is implemented for Ver4. So one can use the API in https://juliacollections.github.io/AbstractTrees.jl/stable/ to manipulate/print the tree structre of Ver4.

There are three different methods to print/visualize the tree structre:

print_tree(ver4::Ver4) or print_tree(bub::Bubble) to print the tree to terminal. This function is provided by AbstractTrees API.

newick(ver4::Ver4) or newick(bub::Bubble) to serilize the tree to a newick format string. You may save the string to a text file, then visualize it with a newick format visualizer application.

showTree(ver4::Ver4) to visualize the tree using the python package ete3. You have to install ete3 properly to use this function.

ExpressionTree.Parquet.newick — Method

convert Bubble tree struct to a string in the newick format

ExpressionTree.Parquet.newick — Method

convert Ver4 tree struct to a string in the newick format

ExpressionTree.Parquet.showTree — Method

showTree(ver4, para::Para; verbose=0, depth=999)

Visualize the diagram tree using ete3 python package

#Arguments

ver4: the 4-vertex diagram tree to visualize

para: parameters

verbose=0: the amount of information to show

depth=999: deepest level of the diagram tree to show