FeynmanDiagram

Documentation for FeynmanDiagram.

FeynmanDiagram.ComputationalGraphs.Graph

FeynmanDiagram.DiagTree.BareGreenNId

FeynmanDiagram.DiagTree.BareHoppingId

FeynmanDiagram.DiagTree.ConnectedGreenNId

FeynmanDiagram.DiagTree.Diagram

FeynmanDiagram.DiagTree.DiagramId

FeynmanDiagram.DiagTree.GreenNId

FeynmanDiagram.DiagTree.PropagatorId

FeynmanDiagram.ExprTree.LoopPool

FeynmanDiagram.ExprTree.Node

FeynmanDiagram.FrontEnds.LabelProduct

FeynmanDiagram.FrontEnds.LoopPool

FeynmanDiagram.Parquet.ParquetBlocks

FeynmanDiagram.QuantumOperators.OperatorProduct

FeynmanDiagram.QuantumOperators.QuantumOperator

Base.:*

Base.adjoint

Base.adjoint

Base.length

Base.show

Base.show

Base.size

Base.size

FeynmanDiagram.ComputationalGraphs.alleq

FeynmanDiagram.ComputationalGraphs.collect_labels

FeynmanDiagram.ComputationalGraphs.eldest

FeynmanDiagram.ComputationalGraphs.external

FeynmanDiagram.ComputationalGraphs.external_labels

FeynmanDiagram.ComputationalGraphs.external_vertex

FeynmanDiagram.ComputationalGraphs.feynman_diagram

FeynmanDiagram.ComputationalGraphs.group

FeynmanDiagram.ComputationalGraphs.haschildren

FeynmanDiagram.ComputationalGraphs.inplace_prod

FeynmanDiagram.ComputationalGraphs.inplace_prod!

FeynmanDiagram.ComputationalGraphs.interaction

FeynmanDiagram.ComputationalGraphs.is_external

FeynmanDiagram.ComputationalGraphs.is_internal

FeynmanDiagram.ComputationalGraphs.isapprox_one

FeynmanDiagram.ComputationalGraphs.isbranch

FeynmanDiagram.ComputationalGraphs.ischain

FeynmanDiagram.ComputationalGraphs.isequiv

FeynmanDiagram.ComputationalGraphs.isfactorless

FeynmanDiagram.ComputationalGraphs.isleaf

FeynmanDiagram.ComputationalGraphs.linear_combination

FeynmanDiagram.ComputationalGraphs.linear_combination

FeynmanDiagram.ComputationalGraphs.merge_prefactors

FeynmanDiagram.ComputationalGraphs.merge_prodchain_subfactors

FeynmanDiagram.ComputationalGraphs.merge_prodchain_subfactors!

FeynmanDiagram.ComputationalGraphs.onechild

FeynmanDiagram.ComputationalGraphs.plot_tree

FeynmanDiagram.ComputationalGraphs.propagator

FeynmanDiagram.ComputationalGraphs.prune_trivial_unary

FeynmanDiagram.ComputationalGraphs.relabel

FeynmanDiagram.ComputationalGraphs.relabel!

FeynmanDiagram.ComputationalGraphs.replace_subgraph

FeynmanDiagram.ComputationalGraphs.replace_subgraph!

FeynmanDiagram.ComputationalGraphs.standardize_labels

FeynmanDiagram.ComputationalGraphs.standardize_labels!

FeynmanDiagram.ComputationalGraphs.vertices

FeynmanDiagram.DiagTree.derivative

FeynmanDiagram.DiagTree.derivative

FeynmanDiagram.DiagTree.plot_tree

FeynmanDiagram.DiagTree.removeDuplicatedLeaves!

FeynmanDiagram.DiagTree.removeHartreeFock!

FeynmanDiagram.DiagTree.removeOneChildParent!

FeynmanDiagram.ExprTree.addnode!

FeynmanDiagram.ExprTree.addpropagator!

FeynmanDiagram.ExprTree.getNode

FeynmanDiagram.ExprTree.getNodeWeight

FeynmanDiagram.ExprTree.showTree

FeynmanDiagram.FrontEnds.index_to_linear

FeynmanDiagram.FrontEnds.linear_to_index

FeynmanDiagram.GV.diagdictGV

FeynmanDiagram.GV.diagdictGV

FeynmanDiagram.GV.eachorder_diag

FeynmanDiagram.GV.leafstates

FeynmanDiagram.GV.read_diagrams

FeynmanDiagram.Parquet.ep_coupling

FeynmanDiagram.Parquet.green

FeynmanDiagram.Parquet.polarization

FeynmanDiagram.Parquet.sigma

FeynmanDiagram.Parquet.vertex3

FeynmanDiagram.Parquet.vertex4

FeynmanDiagram.QuantumOperators.correlator_order

FeynmanDiagram.QuantumOperators.fermionic_annihilation

FeynmanDiagram.QuantumOperators.isannihilation

FeynmanDiagram.QuantumOperators.iscreation

FeynmanDiagram.QuantumOperators.isfermionic

FeynmanDiagram.QuantumOperators.isfermionic

FeynmanDiagram.QuantumOperators.normal_order

FeynmanDiagram.QuantumOperators.parity

FeynmanDiagram.QuantumOperators.parity_old

FeynmanDiagram.innerTauNum

FeynmanDiagram.innerTauNum

Parameters.reconstruct

Parameters.reconstruct

FeynmanDiagram.innerTauNum — Method

function innerTauNum(type, innerLoopNum, interactionTauNum)

internal imaginary-time degrees of freedom for a given diagram type and internal loop number.
For the vertex functions (self-energy, polarization, vertex3, and vertex4), innerTauNum is equivalent to tauNum.
For the Green function, tauNum = innerTauNum + external tauNum

FeynmanDiagram.innerTauNum — Method

function innerTauNum(type::DiagramType, innerLoopNum, interactionTauNum)

internal imaginary-time degrees of freedom for a given diagram type and internal loop number.
For the vertex functions (self-energy, polarization, vertex3, and vertex4), innerTauNum is equivalent to tauNum.
For the Green function, tauNum = innerTauNum + external tauNum

Parameters.reconstruct — Method

Parameters.reconstruct(p::DiagPara; kws...)

Type-stable version of the Parameters.reconstruct

Parameters.reconstruct — Method

Parameters.reconstruct(p::DiagramPara; kws...)

Type-stable version of the Parameters.reconstruct

Library Outline

Quantum operator and its product

API

Computational graph for general feynman diagrams

API

Front-end translates a source code into a computational graph

API

GV expansion to a computational graph

API

Parquet Algorithm to Build Diagrams

API

Usage

Diagrams as an AbstractTree

API

Diagrams as an Expression Tree

API