GV expansion to a computational graph

API

FeynmanDiagram.GV.PolarDiagrams — Function

function PolarDiagrams(type::Symbol, MaxOrder::Int, has_counterterm::Bool=false, dim::Int=3)

Generates a `Graph`: the `dim`-dimensional polarization diagrams with static interactions in a given `type`, to a given maximum order `MaxOrder`, with switchable couterterms. 
Generates fermionic/bosonic `LabelProduct`: `fermi_labelProd`/`bose_labelProd` for this `Graph`.

Arguments:

type (Symbol): The type of the polarization diagrams, either :spin or :charge.

Maxorder (Int): The maximum actual order of the diagrams.

has_counterterm (Bool, optional): false for G0W0, true for GW with interaction and self-energy counterterms.

dim (Int, optional): The dimension of the system (defaults to 3).

Returns

A tuple (diagrams, fermi_labelProd, bose_labelProd) where

diagrams is a Graph objects representing the diagrams,

fermi_labelProd is a LabelProduct object containing the labels for the fermionic G objects in the diagrams,

bose_labelProd is a LabelProduct object containing the labels for the bosonic W objects in the diagrams.

FeynmanDiagram.GV.PolarEachOrder — Function

function PolarEachOrder(type::Symbol, order::Int, VerOrder::Int=0, GOrder::Int=0; loopPool::Union{LoopPool,Nothing}=nothing,
    tau_labels::Union{Nothing,Vector{Int}}=nothing, GTypes::Union{Nothing,Vector{Int}}=nothing, VTypes::Union{Nothing,Vector{Int}}=nothing)

Generates a `Graph`: the polarization diagrams with static interactions of a given order, where the actual order of diagrams equals to `order + VerOrder + 2 * GOrder`.
Generates fermionic/bosonic `LabelProduct`: `fermi_labelProd`/`bose_labelProd` with inputs `tau_labels`, `GTypes`/`VTypes`, and updated `loopPool`.

Arguments:

type (Symbol): The type of the diagrams, either :spin or :charge.

order (Int): The order of the diagrams without counterterms.

VerOrder (Int, optional): The order of interaction counterterms (defaults to 0).

GOrder (Int, optional): The order of self-energy counterterms (defaults to 0).

dim (Int, optional): The dimension of the system (defaults to 3).

loopPool (Union{LoopPool,Nothing}=nothing, optional): The initial pool of loop momenta (defaults to nothing).

tau_labels(Union{Nothing, Vector{Int}}, optional): The labels for the discrete time of each vertex. (defaults to nothing).

GTypes: The types of fermion propagators G in the diagrams (defaults to collect(0:GOrder)).

VTypes: The types of boson static interaction V in the diagrams (defaults to collect(0:VerOrder)).

Returns

A tuple (diagrams, fermi_labelProd, bose_labelProd) where

diagrams is a Graph objects representing the diagrams,

fermi_labelProd is a LabelProduct object containing the labels for the fermionic G objects in the diagrams,

bose_labelProd is a LabelProduct object containing the labels for the bosonic W objects in the diagrams.

FeynmanDiagram.GV.read_diagrams — Method

function read_diagrams(filename::AbstractString; loopPool::Union{LoopPool,Nothing}=nothing,
    dim::Int=3, tau_labels::Union{Nothing,Vector{Int}}=nothing, GTypes=[0, 1], VTypes=[0, 1, 2],
    keywords::Vector{String}=["Polarization", "DiagNum", "Order", "GNum", "Ver4Num", "LoopNum", "ExtLoopIndex",
        "DummyLoopIndex", "TauNum", "ExtTauIndex", "DummyTauIndex"])

Reads a GV_diagrams file and returns Graph of diagrams in this file 
and the corresponding `LabelProduct` objects, which are used to keep track of QuantumOperator.label.

Arguments:

filename (AbstractString): The path to the file containing the diagrams.

loopPool (Union{LoopPool,Nothing}): An optional LoopPool object. If not provided, a new one will be created.

dim (Int): The dimension of the system, used to initialize the LoopPool object. Default is 3.

tau_labels (Union{Nothing,Vector{Int}}): The labels for the Tau objects in the diagrams. If not provided, they will be set to the integers from 1 to tauNum.

GTypes (Vector{Int}): The labels for the fermionic G objects in the diagrams. Default is [0, 1].

VTypes (Vector{Int}): The labels for the bosonic V objects in the diagrams. Default is [0, 1, 2].

keywords (Vector{String}): A set of keywords used to extract information from the file. Default is ["Polarization", "DiagNum", "Order", "GNum", "Ver4Num", "LoopNum", "ExtLoopIndex", "DummyLoopIndex", "TauNum", "ExtTauIndex", "DummyTauIndex"].

Returns

A tuple (diagrams, fermi_labelProd, bose_labelProd) where

diagrams is a Graph object representing the diagrams,

fermi_labelProd is a LabelProduct object containing the labels for the fermionic G objects in the diagrams,

bose_labelProd is a LabelProduct object containing the labels for the bosonic W objects in the diagrams.