Diagrams as an Expression Tree

API

FeynmanDiagram.ExprTree.CachedPool — Type

struct CachedPool{O,T}

    Use this pool to host the objects that are heavy to evaluate so that one wants to cache their status.
    The user should defines a compare

Members

name::Symbol : name of the pool

object::O : object

current::T : current status

new::T : the new status wants to assign later

version::Int128 : the current version

excited::Bool : if set to excited, then the current status needs to be replaced with the new status

FeynmanDiagram.ExprTree.Diagrams — Type

mutable struct Diagrams{V,P,PARA,F,W}

Diagram Object represents a set of Feynman diagrams in a tree (forest) structure

Members

basisPool::V : Tuple of pools of cached basis in a format of (BasisPool1, BasisPool2, ...)

propagatorPool::P : Tuple of pools of cached propagators in a format of (PropagatorPool1, PropagatorPool2, ...)

nodePool::Pool{Node{PARA,F},W} : Pool of the nodes in the diagram tree

root::Vector{Int} : indices of the cached nodes that are the root(s) of the diagram tree. Each element corresponds to one root.

FeynmanDiagram.ExprTree.LoopPool — Type

struct LoopPool{T}

Pool of loop basis. Each loop basis corresponds to a loop variable.
A loop variable is a linear combination of N independent loops. The combination coefficients is what we call a loop basis.
For example, if a loop is a momentum K, then

varibale_i = K_1*basis[1, i] + K_2*basis[2, i] + K_3*basis[3, i] + ...

Members

name::Symbol : name of the pool

dim::Int : dimension of a loop variable (for example, the dimension of a momentum-frequency loop variable is (d+1) where d is the spatial dimension)

N::Int : number of independent loops (dimension of loop basis)

basis::Matrix{T} : Matrix of (N x Nb) that stores the loop basis, where Nb is the number of loop basis (or number of loop variables).

current::Matrix{T} : Matrix of (dim x Nb) that stores the loop variables, where Nb is the number of loop basis (or number of loop variables).

FeynmanDiagram.ExprTree.Node — Type

mutable struct Node{PARA,F}

Node Object, which is the building block of the diagram tree. Each node is a collection of CACHED proapgator objects and other child CACHED node objects

Members

para::PARA : user-defined parameters, which will be used to evaluate the factor and the weight of the node (e.g., if the node represents a vertex function, then the parameter may be the momentum basis of the external legs)

operation::Int : #1: multiply, 2: add, ...

factor::F : additional factor of the node

components::Vector{Vector{Int}} : Index to the cached propagators stored in certain pools. Each Vector{Int} is for one kind of propagator.

childNodes::Vector{Int} : Indices to the cached nodes stored in certain pool. They are the child of the current node in the diagram tree.

parent::Int : Index to the cached nodes which is the parent of the current node.

FeynmanDiagram.ExprTree.Propagator — Type

mutable struct Propagator{PARA,F}

 Object of all kinds of propagators (many-body Green's functions, interactions, vertex functions, ...)

Members

para::PARA : User-defined parameters, which will be used to evaluate the factor and the weight of the propagator

order::Int : The order of the propagators (for example, in a Feynman diagrammatic expansion, the one-body Green's function usually has order zero, and an interaction usually has order one)

factor::F : Additional factor of the propagator

basis::Vector{Int} : Index to the cached basis stored in certain pool. They are essentail for the weight evaluation of the propagator (the first index for the first basis type, the second index for the second basis type, etc.).

FeynmanDiagram.ExprTree.addNode! — Function

function addNode!(diag::Diagrams, operator, name, factor = 1.0; propagator = nothing, child = [], parent = 0, para = nothing)

Add a node into the diagram tree.

Arguments

diag::Diagrams : Diagram tree.

operator::Int : #1: multiply, 2: add, ...

name : name of the node

factor = 1.0 : Factor of the node

propagator = nothing : Index to the cached propagators stored in certain pools. It should be in the format of Vector{Vector{Int}}, where each Vector{Int} is for one kind of propagator.

child = [] : Indices to the cached nodes stored in certain pool. They are the child of the current node in the diagram tree. It should be in the format of Vector{Int}.

para = nothing : Additional paramenter required to evaluate the node. Set to nothing by default.

FeynmanDiagram.ExprTree.addNodeByName! — Function

function addNodeByName(diag::Diagrams, operator, name = :none; child = [], factor = 1.0, parent = 0, para = nothing, kwargs...)

An alternative way to add node. Instead of providing the components vector, one specify the propagators in kwargs.

Arguments

diag::Diagrams : Diagram tree.

operator::Int : #1: multiply, 2: add, ...

name : name of the node

factor = 1.0 : Factor of the node

child = [] : Indices to the cached nodes stored in certain pool. They are the child of the current node in the diagram tree. It should be in the format of Vector{Int}.

para = nothing : Additional paramenter required to evaluate the node. Set to nothing by default.

kwargs... : Accept arbitrary pair of parameters as: PropagatorPoolName = [index of the propagator1, index of the propagator2, ...]

FeynmanDiagram.ExprTree.addPropagator! — Function

function addPropagator!(diag::Diagrams, index::Int, order::Int, name, factor = 1; site = [], loop = nothing, para = nothing)

Add a propagator into the diagram tree.

Arguments

diag::Diagrams : Diagram tree.

index::Int : Index of the propagator in the diagram tree.

order::Int : Order of the propagator.

name = :none : name of the propagator.

factor = 1 : Factor of the propagator.

site = [] : site basis (e.g, time and space coordinate) of the propagator.

loop = nothing : loop basis (e.g, momentum and frequency) of the propagator.

para = nothing : Additional paramenter required to evaluate the propagator.

FeynmanDiagram.ExprTree.addPropagator! — Function

function addPropagator!(diag::Diagrams, poolName::Symbol, order::Int, name, factor = 1; site = [], loop = nothing, para = nothing)

Add a propagator into the diagram tree referenced by the pool name.

Arguments

diag::Diagrams : Diagram tree.

poolName::Symbol : name of the propagator pool

order::Int : Order of the propagator.

name = :none : name of the propagator.

factor = 1 : Factor of the propagator.

site = [] : site basis (e.g, time and space coordinate) of the propagator.

loop = nothing : loop basis (e.g, momentum and frequency) of the propagator.

para = nothing : Additional paramenter required to evaluate the propagator.

FeynmanDiagram.ExprTree.getNode — Method

function getNode(diag::Diagrams, nidx::Int)

get Node in the diag with the index nidx.

FeynmanDiagram.ExprTree.getNodeWeight — Method

function getNodeWeight(diag::Diagrams, nidx::Int)

get Node weight in the diag with the index nidx.

FeynmanDiagram.ExprTree.showTree — Method

showTree(diag::Diagrams, _root = diag.root[end]; verbose = 0, depth = 999)

Visualize the diagram tree using ete3 python package

#Arguments

diag: the Diagrams struct to visualize

_root: the index of the root node to visualize

verbose=0: the amount of information to show

depth=999: deepest level of the diagram tree to show