Diagram Tree Builder based on the Parquet Algorithm

ExpressionTree.Builder.Parquet.IdxMap — Type

struct IdxMap(lv, rv, G0, Gx, ver)

Map left vertex Tpair[lv], right vertex Tpair[rv], the shared Green's function G0[G0] and the channel specific Green's function Gx[Gx] to the top level 4-vertex Tpair[ver]

ExpressionTree.Builder.Parquet.Para — Type

struct Para

Parameters to generate diagrams using Parquet algorithm

Arguments:

loopDim : spatial dimension of the momentum variable

internalLoopNum : number of internal loops

externalLoopNum : number of external loops

chan : channels to be included, [Parquet.T, Parquet.U, Parquet.S, ...]

F : channels of left sub-vertex for the particle-hole and particle-hole-exchange bubbles

V : channels of left sub-vertex for the particle-particle bubble

interactionTauNum : how many τ variables in the interaction function (1 for instantaneous interactoion)

spin : 1 for spinless particle, 2 for spin-1/2 particle

ExpressionTree.Builder.Parquet.Ver4 — Type

function Ver4{W}(para::Para, loopNum, tidx = 1; chan = para.chan, level = 1, id = [1,]) where {W}

Generate 4-vertex diagrams using Parquet Algorithm

#Arguments

para: parameters

loopNum: momentum loop degrees of freedom of the 4-vertex diagrams

tidx: the first τ variable index. It will be the τ variable of the left incoming electron for all 4-vertex diagrams

chan: list of channels of the current 4-vertex. If not specified, it is set to be para.chan

interactionTauNum: list of possible τ degrees of freedom of the bare interaction 0, 2, or 4

level: level in the diagram tree

id: the first element will be used as the id of the Ver4. All nodes in the tree will be labeled in preorder depth-first search

#Remark:

The argument chan and para.chan are different. The former is the channels of current 4-vertex, while the later is the channels of the sub-vertices

AbstractTrees interface is implemented for Ver4. So one can use the API in https://juliacollections.github.io/AbstractTrees.jl/stable/ to manipulate/print the tree structre of Ver4.

There are three different methods to print/visualize the tree structre:

print_tree(ver4::Ver4) or print_tree(bub::Bubble) to print the tree to terminal. This function is provided by AbstractTrees API.

newick(ver4::Ver4) or newick(bub::Bubble) to serilize the tree to a newick format string. You may save the string to a text file, then visualize it with a newick format visualizer application.

showTree(ver4::Ver4) to visualize the tree using the python package ete3. You have to install ete3 properly to use this function.

ExpressionTree.Builder.Parquet.build — Method

function build(weightType::DataType, para::Para, LegK)

build DiagTree for the one-particle-irreducible 4-point vertex function using the parquet algorithm

Arguments:

weightType : type of the weight of the propagators and the vertex functions

para : parameters to generate the diagram tree

LegK : momentum basis of external legs, only three of them are expected: [left in, left out, right in], the dimension of each legK is called loopBasis dimension.

ExpressionTree.Builder.Parquet.newick — Method

convert Bubble tree struct to a string in the newick format

ExpressionTree.Builder.Parquet.newick — Method

convert Ver4 tree struct to a string in the newick format

ExpressionTree.Builder.Parquet.showTree — Method

showTree(ver4, para::Para; verbose=0, depth=999)

Visualize the diagram tree using ete3 python package

#Arguments

ver4: the 4-vertex diagram tree to visualize

para: parameters

verbose=0: the amount of information to show

depth=999: deepest level of the diagram tree to show