Parquet Algorithm to Build Diagrams

API

FeynmanDiagram.Parquet.ParquetBlocks — Type

struct ParquetBlocks

The channels of the left and right sub-vertex4 of a bubble diagram in the parquet equation

#Members

phi : channels of left sub-vertex for the particle-hole and particle-hole-exchange bubbles

ppi : channels of left sub-vertex for the particle-particle bubble

Γ4 : channels of right sub-vertex of all channels

FeynmanDiagram.Parquet.ep_coupling — Method

function ep_coupling(para::DiagPara{W};
    extK=[DiagTree.getK(para.totalLoopNum, 1), DiagTree.getK(para.totalLoopNum, 2), DiagTree.getK(para.totalLoopNum, 3)],
    channels::AbstractVector=[PHr, PHEr, PPr, Alli],
    subdiagram=false,
    name=:none, resetuid=false,
    blocks::ParquetBlocks=ParquetBlocks()
) where {W}

Generate electron-phonon 4-vertex diagrams using Parquet Algorithm. The right incoming Tau will be set to the last Tau for all diagrams | | Γ3 –––-| | |

Arguments

para : parameters. It should provide internalLoopNum, interactionTauNum, firstTauIdx

extK : basis of external loops as a vector [left in, left out, right in, right out].

channels : vector of channels in the left Γ3 diagrams.

subdiagram : a sub-vertex or not

name : name of the vertex

resetuid : restart uid count from 1

blocks : building blocks of the Parquet equation. See the struct ParquetBlocks for more details.

Output

A DataFrame with fields :response, :extT, :diagram, :hash.

Output

A DataFrame with fields :response, :type, :extT, :diagram, :hash

FeynmanDiagram.Parquet.green — Method

green(para::DiagPara, extK = DiagTree.getK(para.totalLoopNum, 1), extT = para.hasTau ? (1, 2) : (0, 0), subdiagram = false;
    name = :G, resetuid = false, blocks::ParquetBlocks=ParquetBlocks())

Build composite Green's function. By definition, para.firstTauIdx is the first Tau index of the left most self-energy subdiagram.

Arguments

para : parameters. It should provide internalLoopNum, interactionTauNum, firstTauIdx

extK : basis of external loop.

extT: [Tau index of the left leg, Tau index of the right leg]

subdiagram : a sub-vertex or not

name : name of the diagram

resetuid : restart uid count from 1

blocks : building blocks of the Parquet equation. See the struct ParquetBlocks for more details.

Output

A Diagram object or nothing if the Green's function is illegal.

FeynmanDiagram.Parquet.polarization — Method

function polarization(para, extK = DiagTree.getK(para.totalLoopNum, 1), subdiagram = false; name = :Π, resetuid = false, blocks::ParquetBlocks=ParquetBlocks())

Generate polarization diagrams using Parquet Algorithm.

Arguments

para : parameters. It should provide internalLoopNum, interactionTauNum, firstTauIdx

extK : basis of external loop.

subdiagram : a sub-vertex or not

name : name of the vertex

resetuid : restart uid count from 1

blocks : building blocks of the Parquet equation. See the struct ParquetBlocks for more details.

Output

A DataFrame with fields :response, :diagram, :hash.

All polarization share the same external Tau index. With imaginary-time variables, they are extT = (para.firstTauIdx, para.firstTauIdx+1)

FeynmanDiagram.Parquet.sigma — Method

function sigma(para, extK = DiagTree.getK(para.totalLoopNum, 1), subdiagram = false; name = :Σ, resetuid = false, blocks::ParquetBlocks=ParquetBlocks())

Build sigma diagram. When sigma is created as a subdiagram, then no Fock diagram is generated if para.filter contains NoFock, and no sigma diagram is generated if para.filter contains Girreducible

Arguments

para : parameters. It should provide internalLoopNum, interactionTauNum, firstTauIdx

extK : basis of external loop.

subdiagram : a sub-vertex or not

name : name of the diagram

resetuid : restart uid count from 1

blocks : building blocks of the Parquet equation. See the struct ParquetBlocks for more details.

Output

A DataFrame with fields :type, :extT, :diagram, :hash

All sigma share the same incoming Tau index, but not the outgoing one

FeynmanDiagram.Parquet.vertex3 — Method

function vertex3(para, extK = [DiagTree.getK(para.totalLoopNum, 1), DiagTree.getK(para.totalLoopNum, 2)],
    subdiagram = false; name = :Γ3, chan = [PHr, PHEr, PPr, Alli], resetuid = false, 
    blocks::ParquetBlocks=ParquetBlocks()
    )

Generate 3-vertex diagrams using Parquet Algorithm. With imaginary-time variables, all vertex3 generated has the same bosonic Tidx $extT[1]=para.firstTauIdx$ and the incoming fermionic Tidx $extT[2]=para.firstTauIdx+1$.

#Arguments

para : parameters. It should provide internalLoopNum, interactionTauNum, firstTauIdx

extK : basis of external loops as a vector [bosonic leg (out), fermionic in, fermionic out], extK[1] = extK[2] - extK[3].

subdiagram : a sub-vertex or not

name : name of the vertex

chan : vector of channels of the current 4-vertex.

resetuid : restart uid count from 1

blocks : building blocks of the Parquet equation. See the struct ParquetBlocks for more details.

Output

A DataFrame with fields :response, :extT, :diagram, :hash.

FeynmanDiagram.Parquet.vertex4 — Method

vertex4(para::DiagPara,
    extK = [DiagTree.getK(para.totalLoopNum, 1), DiagTree.getK(para.totalLoopNum, 2), DiagTree.getK(para.totalLoopNum, 3)],
    chan::AbstractVector = [PHr, PHEr, PPr, Alli],
    subdiagram = false;
    level = 1, name = :none, resetuid = false,
    blocks::ParquetBlocks=ParquetBlocks(),
    blockstoplevel::ParquetBlocks=blocks
    )

Generate 4-vertex diagrams using Parquet Algorithm

Arguments

para : parameters. It should provide internalLoopNum, interactionTauNum, firstTauIdx

extK : basis of external loops as a vector [left in, left out, right in, right out].

chan : vector of channels of the current 4-vertex.

subdiagram : a sub-vertex or not

name : name of the vertex

level : level in the diagram tree

resetuid : restart uid count from 1

blocks : building blocks of the Parquet equation. See the struct ParquetBlocks for more details.

blockstoplevel : building blocks of the Parquet equation at the toplevel. See the struct ParquetBlocks for more details.

Output

A DataFrame with fields :response, :type, :extT, :diagram, :hash

Usage

julia> using FeynmanDiagram
julia> para = DiagPara(type = Ver4Diag, innerLoopNum = 1, hasTau = true);ERROR: MethodError: no method matching DiagPara(::FeynmanDiagram.DiagramType, ::Int64, ::Bool, ::Int64, ::Int64, ::Vector{Interaction}, ::Int64, ::Int64, ::Bool, ::Int64, ::Int64, ::Vector{FeynmanDiagram.Filter}, ::Vector{Any}, ::Type{Nothing})
julia> Parquet.vertex4(para)ERROR: UndefVarError: `para` not defined
julia> para = DiagPara(type = Ver3Diag, innerLoopNum = 1, hasTau = true);ERROR: MethodError: no method matching DiagPara(::FeynmanDiagram.DiagramType, ::Int64, ::Bool, ::Int64, ::Int64, ::Vector{Interaction}, ::Int64, ::Int64, ::Bool, ::Int64, ::Int64, ::Vector{FeynmanDiagram.Filter}, ::Vector{Any}, ::Type{Nothing})
julia> Parquet.vertex3(para)ERROR: UndefVarError: `para` not defined
julia> para = DiagPara(type = SigmaDiag, innerLoopNum = 1, hasTau = true);ERROR: MethodError: no method matching DiagPara(::FeynmanDiagram.DiagramType, ::Int64, ::Bool, ::Int64, ::Int64, ::Vector{Interaction}, ::Int64, ::Int64, ::Bool, ::Int64, ::Int64, ::Vector{FeynmanDiagram.Filter}, ::Vector{Any}, ::Type{Nothing})
julia> Parquet.sigma(para)ERROR: UndefVarError: `para` not defined
julia> para = DiagPara(type = PolarDiag, innerLoopNum = 1, hasTau = true);ERROR: MethodError: no method matching DiagPara(::FeynmanDiagram.DiagramType, ::Int64, ::Bool, ::Int64, ::Int64, ::Vector{Interaction}, ::Int64, ::Int64, ::Bool, ::Int64, ::Int64, ::Vector{FeynmanDiagram.Filter}, ::Vector{Any}, ::Type{Nothing})
julia> Parquet.polarization(para)ERROR: UndefVarError: `para` not defined