Parquet Algorithm to Build Diagrams

API

FeynmanDiagram.Parquet.ParquetBlocks — Type

struct ParquetBlocks

The channels of the left and right sub-vertex4 of a bubble diagram in the parquet equation

#Members

phi : channels of left sub-vertex for the particle-hole and particle-hole-exchange bubbles

ppi : channels of left sub-vertex for the particle-particle bubble

Γ4 : channels of right sub-vertex of all channels

FeynmanDiagram.Parquet.green — Function

green(para, extK = DiagTree.getK(para.totalLoopNum, 1), extT = para.hasTau ? (1, 2) : (0, 0), subdiagram = false;
    name = :G, resetuid = false)

Build composite Green's function.
By definition, para.firstTauIdx is the first Tau index of the left most self-energy subdiagram.

Arguments

para : parameters. It should provide internalLoopNum, interactionTauNum, firstTauIdx

extK : basis of external loop.

extT: [Tau index of the left leg, Tau index of the right leg]

subdiagram : a sub-vertex or not

name : name of the diagram

resetuid : restart uid count from 1

Output

A Diagram object or nothing if the Green's function is illegal.

FeynmanDiagram.Parquet.polarization — Function

function polarization(para, extK = DiagTree.getK(para.totalLoopNum, 1), subdiagram = false; name = :Π, resetuid = false)

Generate polarization diagrams using Parquet Algorithm.

Arguments

para : parameters. It should provide internalLoopNum, interactionTauNum, firstTauIdx

extK : basis of external loop.

subdiagram : a sub-vertex or not

name : name of the vertex

resetuid : restart uid count from 1

Output

A DataFrame with fields :response, :diagram, :hash.

All polarization share the same external Tau index. With imaginary-time variables, they are extT = (para.firstTauIdx, para.firstTauIdx+1)

FeynmanDiagram.Parquet.sigma — Function

function sigma(para, extK = DiagTree.getK(para.totalLoopNum, 1), subdiagram = false; name = :Σ, resetuid = false)

Build sigma diagram. 
When sigma is created as a subdiagram, then no Fock diagram is generated if para.filter contains NoFock, and no sigma diagram is generated if para.filter contains Girreducible

Arguments

para : parameters. It should provide internalLoopNum, interactionTauNum, firstTauIdx

extK : basis of external loop.

subdiagram : a sub-vertex or not

name : name of the diagram

resetuid : restart uid count from 1

Output

A DataFrame with fields :type, :extT, :diagram, :hash

All sigma share the same incoming Tau index, but not the outgoing one

FeynmanDiagram.Parquet.vertex3 — Function

function vertex3(para, extK = [DiagTree.getK(para.totalLoopNum, 1), DiagTree.getK(para.totalLoopNum, 2)],
    subdiagram = false; name = :Γ3, chan = [PHr, PHEr, PPr, Alli], resetuid = false)

Generate 3-vertex diagrams using Parquet Algorithm.
With imaginary-time variables, all vertex3 generated has the same bosonic Tidx ``extT[1]=para.firstTauIdx`` and the incoming fermionic Tidx ``extT[2]=para.firstTauIdx+1``.

#Arguments

para : parameters. It should provide internalLoopNum, interactionTauNum, firstTauIdx

extK : basis of external loops as a vector [bosonic leg, fermionic in, fermionic out].

subdiagram : a sub-vertex or not

name : name of the vertex

chan : vector of channels of the current 4-vertex.

resetuid : restart uid count from 1

Output

A DataFrame with fields :response, :extT, :diagram, :hash.

FeynmanDiagram.Parquet.vertex4 — Function

vertex4(para::GenericPara,
    extK = [DiagTree.getK(para.totalLoopNum, 1), DiagTree.getK(para.totalLoopNum, 2), DiagTree.getK(para.totalLoopNum, 3)],
    chan::AbstractVector = [PHr, PHEr, PPr, Alli],
    subdiagram = false;
    level = 1, name = :none, resetuid = false,
    phi_toplevel = ParquetBlocks().phi, ppi_toplevel = ParquetBlocks().ppi, Γ4_toplevel = ParquetBlocks().Γ4)

Generate 4-vertex diagrams using Parquet Algorithm

Arguments

para : parameters. It should provide internalLoopNum, interactionTauNum, firstTauIdx

extK : basis of external loops as a vector [left in, left out, right in, right out].

chan : vector of channels of the current 4-vertex.

subdiagram : a sub-vertex or not

name : name of the vertex

level : level in the diagram tree

resetuid : restart uid count from 1

phi_toplevel : channels of left sub-vertex for the particle-hole and particle-hole-exchange of the bubble at level one.

ppi_toplevel : channels of left sub-vertex for the particle-particle bubble at level one

Γ4_toplevel : channels of right sub-vertex for all all bubbles at level one

Output

A DataFrame with fields :response, :type, :extT, :diagram, :hash

Usage

julia> using FeynmanDiagram
julia> para = GenericPara(diagType = Ver4Diag, innerLoopNum = 1, hasTau = true);
julia> Parquet.vertex4(para)6×5 DataFrame
 Row │ response  type     extT          diagram                            has ⋯
     │ Any       Any      Any           Diagram…                           Int ⋯
─────┼──────────────────────────────────────────────────────────────────────────
   1 │ UpUp      Dynamic  (1, 1, 2, 2)  ↑↑Dyn#[0, 0],t(1, 1, 2, 2)=0.0=⨁…      ⋯
   2 │ UpUp      Dynamic  (1, 2, 1, 2)  ↑↑x↑↑ → PPr, PPr ↑↑Dyn#[0, 0],t(…
   3 │ UpUp      Dynamic  (1, 2, 2, 1)  ↑↑Dyn#[0, 0],t(1, 2, 2, 1)=0.0=⨁…
   4 │ UpDown    Dynamic  (1, 1, 2, 2)  ↑↓Dyn#[0, 0],t(1, 1, 2, 2)=0.0=⨁…
   5 │ UpDown    Dynamic  (1, 2, 1, 2)  ↑↓Dyn#[0, 0],t(1, 2, 1, 2)=0.0=⨁…      ⋯
   6 │ UpDown    Dynamic  (1, 2, 2, 1)  ↑↓Dyn#[0, 0],t(1, 2, 2, 1)=0.0=⨁…
                                                                1 column omitted
julia> para = GenericPara(diagType = Ver3Diag, innerLoopNum = 1, hasTau = true);
julia> Parquet.vertex3(para)2×4 DataFrame
 Row │ response  extT       diagram                            hash
     │ Response  Tuple…     Diagram…                           Int64
─────┼───────────────────────────────────────────────────────────────
   1 │ UpUp      (1, 2, 2)  Γ3 ↑↑#[0, 0],t(1, 2, 2)=0.0=4.03…     95
   2 │ UpDown    (1, 2, 2)  Γ3 ↑↓#[0, 0],t(1, 2, 2)=0.0=4.03…     98
julia> para = GenericPara(diagType = SigmaDiag, innerLoopNum = 1, hasTau = true);
julia> Parquet.sigma(para)1×4 DataFrame
 Row │ type       extT    diagram                            hash
     │ Analytic…  Tuple…  Diagram…                           Int64
─────┼─────────────────────────────────────────────────────────────
   1 │ Instant    (1, 1)  Σ Ins#[0, 0], t(1, 1)=0.0=4.031e…    107
julia> para = GenericPara(diagType = PolarDiag, innerLoopNum = 1, hasTau = true);
julia> Parquet.polarization(para)1×4 DataFrame
 Row │ response  extT    diagram                            hash
     │ Response  Tuple…  Diagram…                           Int64
─────┼────────────────────────────────────────────────────────────
   1 │ UpUp      (1, 2)  Π ↑↑#[0, 0], k[1.0, 0.0], t(1, 2…    110